# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Connected Care API
  version: 1.0.0
  description: |
    The Connected Care API tier: device telemetry ingestion, the insight engine,
    breed reference data, containment/safety status, Claude-powered agents, and
    the demo presentation tier — the same contracts the production mobile app
    would build against.

    ## Conventions
    - **Auth**: sign in via `POST /v1/auth/login` to receive a signed token
      (HS256 JWT) carrying the owner's household claim, then send it as
      `Authorization: Bearer <token>` on every endpoint except `GET /v1/health`,
      `POST /v1/auth/login`, `GET /v1/auth/demo-identities`, and `/docs`.
      Invalid, expired, or missing tokens → 401. *Production note: a real IdP
      (OIDC) replaces the login route only — the claim contract and enforcement
      are unchanged.*
    - **Envelopes**: success responses wrap payloads as `{"data": ...}`; errors
      are `{"error": {"code", "message", "details?"}}`. Validation failures
      return `code: "validation_error"` with per-field `details`.
    - **Timestamps** are ISO-8601 UTC strings.
    - **Tenancy**: `household` is the tenant boundary, and it is enforced on
      every tenant-scoped route: the resource's household (resolved via pet or
      device where needed) must match the token's household claim, else
      403 `forbidden`. One token can never read another household's data.
    - **Scoring**: the demo scores telemetry synchronously on ingestion for
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/households/{householdId}/stream:
    get:
      tags: [Realtime]
      summary: Server-Sent Events stream of new insights
      security: []
      description: |
        `text/event-stream`. Frames: `event: insight` with the full Insight JSON
        as `data`; comment heartbeats (`: ping`) every 15 s.

        `EventSource` cannot set an `Authorization` header, so the session
        token travels as the `token` query parameter instead. The server
        verifies it and streams **only the token's own household** — a token
        for another household gets 403 regardless of the path id.

        ```js
        const es = new EventSource(`/v1/households/${me.household_id}/stream?token=${token}`);
        es.addEventListener('insight', (e) => console.log(JSON.parse(e.data)));
        ```

        OpenAPI cannot fully model SSE; this entry documents the contract
        informally.
      parameters:
        - { $ref: '#/components/parameters/householdId' }
        - name: token
          in: query
          required: true
          schema: { type: string }
          description: The signed session token (same value as the bearer token elsewhere).
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '200':
          description: An open SSE stream.
          content:
            text/event-stream:
              schema: { type: string }
              example: |
                : connected

                event: insight
                data: {"id":"...","urgency":"urgent","summary":"..."}
  /v1/households/{householdId}/briefing:
    get:
      tags: [Engagement]
      summary: Daily briefing — narrated overnight digest
      description: >-
        Facts (per-pet sleep/walks, fence activity, equipment posture, worst
        active alert) computed live; narration by Claude when available,
        deterministic template from the same facts otherwise (`mode` says which).
      parameters: [{ $ref: '#/components/parameters/householdId' }]
      responses:
        '200':
          description: The briefing.
          content:
            application/json:
              example:
                data:
                  greeting_period: morning
                  summary: The whole crew had a good night — fence quiet, everyone on baseline. 💚
                  facts: [{ icon: '🐾', text: 'Baxter: slept 10.2h, walked 27 min yesterday' }]
                  mode: template
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/households/{householdId}/score:
    get:
      tags: [Engagement]
      summary: Peace-of-Mind score (0–100) with named factors
      description: >-
        Deducted live from active insights, containment posture, and equipment
        readiness — always explainable via `factors`. Bands: ≥90 protected,
        ≥70 good, ≥40 needs_attention, else act_now.
      parameters: [{ $ref: '#/components/parameters/householdId' }]
      responses:
        '200':
          description: The score.
          content:
            application/json:
              example:
                data:
                  score: 92
                  band: protected
                  factors: [{ label: '💚 Every device is online and every trend is on baseline', delta: 0 }]
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/households/{householdId}/milestones:
    get:
      tags: [Engagement]
      summary: Wins worth celebrating, computed from the telemetry record
      parameters: [{ $ref: '#/components/parameters/householdId' }]
      responses:
        '200':
          description: Up to four current wins (walk streaks, breach-free runs, steady hydration).
          content:
            application/json:
              example:
                data:
                  - { icon: '🛡️', title: '21+ days breach-free', detail: 'Nobody has crossed the boundary', pet_id: null }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/households/{householdId}/dealer:
    get:
      tags: [Engagement]
      summary: The household's dealer, with live dispatch state
      description: >-
        Who serves this yard, plus a dispatch timeline (alerted → reviewing →
        followed_up) whenever an emergency insight has been routed to the
        associate queue. Demo note: steps advance deterministically with
        elapsed time — the stand-in for the dealer CRM's status feed.
      parameters: [{ $ref: '#/components/parameters/householdId' }]
      responses:
        '200':
          description: Dealer status (`dealer` null when none serves the household).
          content:
            application/json:
              example:
                data:
                  dealer: { name: Invisible Fence Brand — Chicago North, associate: Dana M., phone: (312) 555-0148 }
                  dispatch: { step: alerted, summary: 'Wrigley crossed the boundary…', routed_at: '2026-07-29T14:02:11Z' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/orders:
    post:
      tags: [Commerce]
      summary: Place a consumable replacement order
      description: >-
        One-tap replenishment from an equipment insight. The sku must match the
        device's consumable (fountain filter, collar battery). Demo note: the
        order is recorded and confirmed with an ETA, never fulfilled;
        production hands off to the commerce platform.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id, sku]
              properties:
                device_id: { type: string }
                sku: { type: string }
            example: { device_id: dev_fountain_9002, sku: filt-std-4pk }
      responses:
        '201':
          description: Order placed.
          content:
            application/json:
              example:
                data: { id: ord_1, sku: filt-std-4pk, label: Fountain filter 4-pack, price_cents: 1299, status: placed, eta_date: '2026-08-01' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/households/{householdId}/orders:
    get:
      tags: [Commerce]
      summary: List the household's orders (newest first)
      parameters: [{ $ref: '#/components/parameters/householdId' }]
      responses:
        '200':
          description: Orders.
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/pets/{petId}/vet-report/shares:
    get:
      tags: [Vet Report]
      summary: Share history with delivery/review stages
      description: >-
        Each share carries a stage — sent → delivered → viewed → reviewed —
        and, once reviewed, a clinic note. Demo note: stages advance
        deterministically with elapsed time since the share; production reads
        the vet portal's receipts and replies.
      parameters: [{ $ref: '#/components/parameters/petId' }]
      responses:
        '200':
          description: Shares, newest first.
          content:
            application/json:
              example:
                data:
                  - { id: shr_1, recipient: Lincoln Park Veterinary Clinic, method: portal, stage: reviewed, clinic_note: 'Trend noted — bring Baxter in if it persists this week. — Dr. Chen' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/pets/{petId}/vet-report:
    get:
      tags: [Vet Report]
      summary: Assemble the live veterinary report
      description: |
        Assembled at request time — never persisted; only shares persist.
        Includes an agent-written professional summary (Claude or deterministic
        template; `summary.mode` says which) under strict guardrails: a
        device-data pattern summary, never a diagnosis. `active_insights`
        excludes dismissed insights (the owner marked them irrelevant) and keeps
        acknowledged ones.
      parameters: [{ $ref: '#/components/parameters/petId' }]
      responses:
        '200':
          description: The report.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/VetReport' } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/pets/{petId}/vet-report/share:
    post:
      tags: [Vet Report]
      summary: Record a digital share of the report
      description: |
        The server snapshots the ids of the currently-active insights into the
        share record. *Production note: v2 integrates real vet portals; the demo
        records and confirms the share without external delivery.*
      parameters: [{ $ref: '#/components/parameters/petId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient, method]
              properties:
                recipient: { type: string, maxLength: 200 }
                method: { type: string, enum: [portal, email, link] }
            example: { recipient: Lincoln Park Veterinary Clinic, method: portal }
      responses:
        '201':
          description: Share recorded.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/VetShare' } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Signed session token from `POST /v1/auth/login` (HS256, 12-hour expiry)
        carrying the owner's household claim. Every tenant-scoped route enforces
        the claim against the requested resource's household.
  parameters:
    householdId:
      name: householdId
      in: path
      required: true
      schema: { type: string }
      example: hh_2291
    petId:
      name: petId
      in: path
      required: true
      schema: { type: string }
      example: pet_5001
...
  responses:
    ValidationError:
      description: Request failed validation; `details` lists per-field failures.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: validation_error
              message: Request body failed validation.
              details: [{ path: device_id, message: device_id is required }]
    Unauthorized:
      description: Missing, malformed, tampered, or expired token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthorized, message: A valid session token is required. } }
    Forbidden:
      description: The token is valid but its household claim does not own the requested resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: forbidden, message: This resource belongs to a different household. } }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: not_found, message: Pet not found. } }
...
  schemas:
    Species:
      type: string
      enum: [dog, cat, other]
    Urgency:
      type: string
      enum: [info, monitor, attention, urgent, emergency]
      description: The five-tier scale that decides how (and whether) an insight reaches the owner.
    InsightType:
      type: string
      enum: [pet_health, pet_safety, equipment]
...
    Metric:
      type: string
      enum: [resting_heart_rate, walk_minutes, water_intake_ml, food_intake_g, sleep_hours, feeder_schedule_drift_min, fountain_flow_rate, collar_battery_pct, boundary_safety, containment_signal]
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    BreedCondition:
      type: object
      properties:
        condition: { type: string, example: mitral valve disease }
        typical_onset: { type: string, example: age 5-7 for half the breed, near all by 10-11 }
        monitorable_today: { type: boolean }
...
    PetBaseline:
      type: object
      properties:
        pet_id: { type: string }
        metric: { $ref: '#/components/schemas/Metric' }
        mean: { type: number, example: 96.2 }
        stdev: { type: number, example: 1.9 }
        window_days: { type: integer, example: 14 }
        sample_count: { type: integer, example: 14 }
        status: { type: string, enum: [insufficient_data, established] }
        last_computed: { type: string, format: date-time }
...
    MetricSeries:
      type: object
      properties:
        metric: { $ref: '#/components/schemas/Metric' }
        points:
          type: array
          items:
            type: object
            properties:
              days_ago: { type: integer, description: Rolling 24h bucket; 0 = the last 24 hours. }
              value: { type: number }
        baseline:
          type: [object, 'null']
          properties:
            mean: { type: number }
            stdev: { type: number }
            status: { type: string }
    Insight:
      type: object
      properties:
        id: { type: string }
        household_id: { type: string }
        pet_id: { type: [string, 'null'] }
        device_id: { type: [string, 'null'] }
        insight_type: { $ref: '#/components/schemas/InsightType' }
        metric: { $ref: '#/components/schemas/Metric' }
        severity_score: { type: number }
        urgency: { $ref: '#/components/schemas/Urgency' }
        summary: { type: string, example: "Baxter's resting heart rate has run 15% above his own baseline for 3 days. Wrigley's readings stay normal over the same period, which rules out a shared household cause." }
        recommended_action: { type: string, example: Schedule a vet visit and mention the resting heart rate trend. }
        generated_at: { type: string, format: date-time }
        acknowledged_status: { type: string, enum: [unseen, seen, acknowledged, dismissed] }
        routed_to_associate: { type: integer, description: SQLite boolean (0/1). }
        narration_mode: { type: string, enum: [claude, template] }
...
    VetShare:
      type: object
      properties:
        id: { type: string }
        household_id: { type: string }
        pet_id: { type: string }
        recipient: { type: string, example: Lincoln Park Veterinary Clinic }
        method: { type: string, enum: [portal, email, link] }
        insight_ids: { type: array, items: { type: string } }
        shared_at: { type: string, format: date-time }
    VetReport:
      type: object
      description: Assembled live at GET time; never persisted.
      properties:
        report_id: { type: string }
        generated_at: { type: string, format: date-time }
        window_days: { type: integer, example: 14 }
        pet:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            species: { $ref: '#/components/schemas/Species' }
            breed_name: { type: [string, 'null'] }
            breed_reference_confidence: { type: string, enum: [high, low] }
            date_of_birth: { type: [string, 'null'], format: date }
            age_years: { type: [integer, 'null'] }
            weight_lbs: { type: [number, 'null'] }
            sex: { type: [string, 'null'] }
        breed:
          type: [object, 'null']
          properties:
            breed_name: { type: [string, 'null'] }
            size_class: { type: string }
            resting_hr_low: { type: integer }
            resting_hr_high: { type: integer }
            common_conditions: { type: array, items: { $ref: '#/components/schemas/BreedCondition' } }
            is_generic_fallback: { type: boolean }
        summary:
          type: object
          properties:
            text: { type: string }
            mode: { type: string, enum: [claude, template] }
        baselines: { type: array, items: { $ref: '#/components/schemas/PetBaseline' } }
        metrics: { type: array, items: { $ref: '#/components/schemas/MetricSeries' } }
        active_insights: { type: array, items: { $ref: '#/components/schemas/Insight' } }
        containment:
          type: [object, 'null']
          properties:
            state: { type: string }
            minutes_since_check_in: { type: [integer, 'null'] }
            battery_pct: { type: [number, 'null'] }
        environment:
          type: [object, 'null']
          properties:
            date: { type: string }
            low_f: { type: [number, 'null'] }
            high_f: { type: [number, 'null'] }
            conditions: { type: [string, 'null'] }
        shares: { type: array, items: { $ref: '#/components/schemas/VetShare' } }
...